+++ xLights/EditAliasesDialog.cpp
@@ -96,7 +96,7 @@ void EditAliasesDialog::ValidateWindow()
     } else {
         ButtonDelete->Disable();
     }
-    if (ListBoxAliases->GetSelection() >= 0 && ListBoxAliases->GetSelection() < ListBoxAliases->GetCount()) {
+    if (ListBoxAliases->GetSelection() >= 0 && ListBoxAliases->GetSelection() < (int)ListBoxAliases->GetCount()) {
         ButtonMoveUp->Enable(true);
         ButtonMoveDown->Enable(true);
     } else {
@@ -176,7 +176,7 @@ void EditAliasesDialog::OnButtonMoveDownClick(wxCommandEvent& event) {
 
     int selectedIndex = ListBoxAliases->GetSelection();
 
-    if (selectedIndex != wxNOT_FOUND && selectedIndex < (ListBoxAliases->GetCount() - 1)) {
+    if (selectedIndex != wxNOT_FOUND && selectedIndex < (int)(ListBoxAliases->GetCount() - 1)) {
         wxString selectedItem = ListBoxAliases->GetString(selectedIndex);
         ListBoxAliases->Delete(selectedIndex);
         ListBoxAliases->Insert(selectedItem, selectedIndex + 1);
