+++ xLights/EffectTreeDialog.cpp
@@ -396,7 +396,7 @@ wxString EffectTreeDialog::ParseLayers(wxString name, wxString settings)
         wxArrayString all_efdata = wxSplit(settings, '\n');
 
         bool cf1 = false;
-        for (int i = 0; i < all_efdata.size(); i++)
+        for (size_t i = 0; i < all_efdata.size(); i++)
         {
             //logger_base.debug("    %d: %s", i, (const char *)all_efdata[i].c_str());
 
@@ -498,7 +498,7 @@ wxString EffectTreeDialog::ParseDuration(wxString name, wxString settings)
         wxArrayString all_efdata = wxSplit(settings, '\n');
 
         bool cf1 = false;
-        for (int i = 0; i < all_efdata.size(); i++)
+        for (size_t i = 0; i < all_efdata.size(); i++)
         {
             //logger_base.debug("    %d: %s", i, (const char *)all_efdata[i].c_str());
